requirements.txt
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
decorator==4.4.2
Flask==2.3.2
google-api-core==2.25.1
google-api-python-client==2.110.0
google-auth==2.25.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
moviepy==1.0.3
mutagen==1.47.0
numpy==2.2.6
oauthlib==3.2.2
Pillow==9.5.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodomex==3.23.0
pyparsing==3.2.3
python-dotenv==1.0.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
tqdm==4.67.1
uritemplate==4.2.0
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
yt-dlp==2025.6.9
